refactor(money control): implement money control to be re-usable
